+++ elements/page-layout.js
@@ -3,77 +3,64 @@ const html = require('choo/html')
 const sf = require('sheetify')
 const panelGroup = require('../components/panel-group')
 
-const mainPrefix = sf`
+const pageLayoutPrefix = sf`
   :host {
     width: 100vw;
     height: 100vh;
-    display: flex;
-    flex-direction: column;
+    display: flexbox;
+  }
+
+  :host > .header {
+    height: 65px;
+    margin: 0;
+    padding: 25px 0 0 100px;
+    font-size: 26px;
+    font-weight: normal;
+    color: #fff;
+    background: #03A9F4;
+    background-image: url('./assets/icons/logo.svg');
+    background-size: 50px;
+    background-position: 45px 5px;
+    background-repeat: no-repeat;    
+  }
+
+  :host > .content {
+    height: calc(100% - 65px)
   }
 `
 
 function pageLayout ({
-  id, // used to get unique instance of panel group
-  header, panels, context,
-  send, onload = _.noop
+  context, id, panels, menu
 }) {
-  const panelGroupHtml = panelGroup.component(...context)(id, {
+  const panelSizes = _.map(panels, 'size')
+  const panelViews = _.map(panels, 'view')
+
+  const contentHtml = panelGroup.component(...context)(id, {
     props: {
-      panelViews: panels
+      panelViews
+    },
+    initialState: {
+      panelSizes: panelSizes
     }
   })
 
   return html`
-    <div class="${mainPrefix}" onload=${onload}>
-      ${getHeaderHtml(header)}
-      ${panelGroupHtml}
+    <div class="${pageLayoutPrefix}">
+      ${getHeaderHtml(menu)}
+      <div class="content">
+        ${contentHtml}
+      </div>
     </div>
   `
 }
 
-const headerPrefix = sf`
-  :host {
-    height: 50px;
-    display: flex;
-    justify-content: space-between;
-    background: #404040;
-  }
-
-  :host > .left-side, :host > .right-side {
-    align-items: center;
-    display: flex;
-  }
-
-  :host > .left-side  {
-    justify-content: flex-start;
-  }
-
-  :host > .right-side {
-    justify-content: flex-end;
-  }
-`
-
-function getHeaderHtml ({ left, right }) {
-  let leftSideHtml, rightSideHtml
-
-  if (left) {
-    leftSideHtml = html`
-      <div class="left-side">${left}</div>
-    `
-  }
-
-  if (right) {
-    rightSideHtml = html`
-      <div class="right-side">${right}</div>
-    `
-  }
-
+function getHeaderHtml (menu) {
   return html`
-    <div class="${headerPrefix}">
-      ${leftSideHtml}
-      ${rightSideHtml}
-    </div>
+    <h1 class="header">
+      Robotopia
+    </h1>
   `
 }
+
 
 module.exports = pageLayout
