+++ pages/tutorial.js
@@ -1,6 +1,6 @@
 const html = require('choo/html')
 const pageLayout = require('../elements/page-layout')
-const gameView = require('../elements/game/index')
+const gameRunnerView = require('../elements/game-runner')
 const tutorialDialogView = require('../elements/tutorial/tutorialDialog')
 const blocklyWidget = require('../elements/blockly')
 const { speedSliderView, playButtonView } = require('../elements/runtime-controls')
@@ -14,32 +14,10 @@ const tutorialView = (state, prev, send) => {
   // we need to check if level param has changed because onLoad will not be triggered if tutorial page
   // has already been loaded
   if (prev !== null && location.params.level !== prev.location.params.level) {
-    initLevel()
+    init()
   }
 
   const instructionHtml = instructionView(game, tutorial)
-
-  const playButtonHtml = playButtonView({
-    isRunning: clock.isRunning,
-    onStart: () => {
-      send('game:loadGameState', { loadState: tutorial.level.game })
-      send('runtime:destroyRobot', { id: 'ROBOT' })
-      send('runtime:createRobot', { id: 'ROBOT', groupId: 1 })
-      send('clock:start')
-    },
-    onStop: () => {
-      send('clock:stop')
-      send('runtime:destroyRobot', { id: 'ROBOT' })
-      send('game:loadGameState', { loadState: tutorial.level.game })
-    }
-  })
-
-  const speedSliderHtml = speedSliderView({
-    min: 100,
-    max: 1000,
-    intervalDuration: clock.intervalDuration,
-    onChange: (value) => send('clock:setIntervalDuration', { intervalDuration: value })
-  })
 
   const blocklyHtml = blocklyView({
     toolbox: tutorial.level && tutorial.level.editor.toolbox,
@@ -50,36 +28,40 @@ const tutorialView = (state, prev, send) => {
     }
   })
 
-  const gameHtml = gameView({
-    state: game,
-    progress: clock.progress
+  const gameRunnerHtml = gameRunnerView({
+    game, clock,
+    onStart: () => {
+      send('game:loadGameState', { loadState: tutorial.level.game })
+      send('runtime:destroyRobot', { id: 'ROBOT' })
+      send('runtime:createRobot', { id: 'ROBOT', groupId: 1 })
+      send('clock:start')
+    },
+    onStop: () => {
+      send('clock:stop')
+      send('runtime:destroyRobot', { id: 'ROBOT' })
+      send('game:loadGameState', { loadState: tutorial.level.game })
+    },
+    onChangeSpeed: (value) => send('clock:setIntervalDuration', { intervalDuration: value })
   })
 
   const pageLayoutHtml = pageLayout({
     id: 'tutorial-page',
     context: [state, prev, send],
-    header: {
-      left: [
-        playButtonHtml,
-        speedSliderHtml
-      ]
-    },
-
     panels: [
-      instructionHtml,
-      blocklyHtml,
-      gameHtml
+      { view: instructionHtml, size: 1 },
+      { view: gameRunnerHtml, size: 2 },
+      { view: blocklyHtml, size: 2 }
     ]
   })
 
   return html`
-    <div onload=${initLevel}>
+    <div onload=${init}>
       ${pageLayoutHtml}
       ${tutorialDialogView(game, tutorial, editor.workspace, send)}
     </div>
   `
 
-  function initLevel () {
+  function init () {
     send('clock:stop')
     send('tutorial:loadLevel', { name: location.params.level })
   }
